# api/http/openapi.yaml
openapi: "3.0.3"
info:
  title: Breakfix API
  version: "1.0.0"
  description: SRE/DevOps interview practice platform
servers:
  - url: http://localhost:9090/api
    description: Local dev
paths:
...
  /challenges/{id}/start:
    post:
      operationId: startChallenge
      summary: Start or resume a challenge environment
      tags: [environments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Environment ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartResponse"
        "404":
          $ref: "#/components/responses/Error"
...
  /challenges/{id}/reset:
    post:
      operationId: resetChallenge
      summary: Reset challenge to initial state
      tags: [environments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetResponse"
        "404":
          $ref: "#/components/responses/Error"
  /challenges/{id}/stop:
    post:
      operationId: stopChallenge
      summary: Stop the current challenge environment
      tags: [environments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopResponse"
        "404":
          $ref: "#/components/responses/Error"
  /challenges/{id}/terminal-ticket:
    post:
      operationId: createTerminalTicket
      summary: Create a one-time ticket for a terminal WebSocket
      tags: [environments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalTicketRequest"
      responses:
        "200":
          description: Short-lived one-time terminal ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalTicketResponse"
        "404":
          $ref: "#/components/responses/Error"
  /challenges/{id}/terminals/{window}:
    delete:
      operationId: closeTerminalWindow
      summary: Close one terminal tab in the current challenge environment
      tags: [environments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: window
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z][a-z0-9-]{0,31}$"
        - name: node
          in: query
          required: false
          schema:
            type: string
            pattern: "^[a-z][a-z0-9-]{0,31}$"
      responses:
        "200":
          description: Terminal window closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalWindowCloseResponse"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    StartResponse:
      type: object
      required: [challenge_title]
      properties:
        challenge_title:
          type: string
    ResetResponse:
      type: object
      required: [challenge_title]
      properties:
        challenge_title:
          type: string
    StopResponse:
      type: object
      required: [stopped, challenge_title]
      properties:
        stopped:
          type: boolean
        challenge_title:
          type: string
    TerminalWindowCloseResponse:
      type: object
      required: [closed]
      properties:
        closed:
          type: boolean
    TerminalTicketRequest:
      type: object
      required: [window]
      properties:
        window:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,31}$"
        node:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,31}$"
    TerminalTicketResponse:
      type: object
      required: [ticket]
      properties:
        ticket:
          type: string
...